Project20/backend/requirements.txt
Django==5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
python-dotenv==1.0.0
pytz==2024.1
